+++ assets/js/modals.js
@@ -1,0 +1,12 @@
+/******/ (() => { // webpackBootstrap
+var __webpack_exports__ = {};
+/*!***********************************!*\
+  !*** ./src/components/modals.tsx ***!
+  \***********************************/
+function initModals(modals) {
+    console.log('ahoj');
+}
+
+/******/ })()
+;
+//# sourceMappingURL=modals.js.map