+++ dev/src/components/modals.tsx
@@ -1,0 +1,7 @@
+
+
+function initModals ( modals:string ):void {
+  
+  console.log( 'ahoj' );
+
+}